+++ src/main/java/ca/sait/vezorla/model/Product.java
@@ -70,10 +70,10 @@ public class Product implements Serializable {
 	
 	@NotNull
 	@Column(name = "price")
-	private BigDecimal price;
+	private long price;
 
 	@Column(name = "old_price")
-	private BigDecimal oldPrice;
+	private long oldPrice;
 
 	@JsonIgnore
 	@OneToMany (mappedBy = "product")
